+++ OpenGLEngine/RenderedObjects/TexturedBoxWithLighting.cs
@@ -59,7 +59,7 @@ namespace OpenGLEngine.RenderedObjects
 					negXdim, Ydim, negZdim, 0,0,-1, r,g,b,a, negUx, Uz,
 					
 			};
-            short[] quadindicedata = {
+            int[] quadindicedata = {
 					0, 1, 2, 
 					0, 2, 3, 
 					4, 5, 6,
